add enviroments
