+++ backend/app.js
@@ -9,7 +9,7 @@ const userRoutes = require('./routes/user');
 const app = express();
 
 mongoose.connect(
-  'mongodb+srv://joey:15Goh66XqDdr9pE4@cluster0-qgyoy.mongodb.net/node-angular?retryWrites=true',
+  'mongodb+srv://joey:' + process.env.MONGO_ATLAS_PW +'@cluster0-qgyoy.mongodb.net/node-angular?retryWrites=true',
   { useNewUrlParser: true },
   )
 .then(() => {
