+++ backend/routes/user.js
@@ -1,67 +1,11 @@
 const express = require('express');
-const bcrypt = require('bcrypt');
-const jwt = require("jsonwebtoken");
 
-const User = require('../models/user');
+const UserController = require('../controllers/user');
 
 const router = express.Router();
 
-router.post("/signup", (req, res, next) => {
-  bcrypt.hash(req.body.password, 10)
-    .then(hash => {
-      const user = new User({
-        email: req.body.email,
-        password: hash,
-      });
-      user.save()
-        .then(result => {
-          res.status(201).json({
-            message: 'User created!',
-            result: result,
-          });
-        })
-        .catch(err => {
-          res.status(500).json({
-              message: 'Invalid auth credentials!',
-          });
-        });
-    });
-});
+router.post("/signup", UserController.createUser);
 
-router.post("/login", (req, res, next) => {
-  let fetchedUser;
-  User.findOne({ email: req.body.email })
-    .then(user => {
-      if (!user) {
-        return res.status(401).json({
-          message: 'Auth failed',
-        });
-      }
-      fetchedUser = user;
-      return bcrypt.compare(req.body.password, user.password);
-    })
-    .then(result => {
-      if (!result) {
-        return res.status(401).json({
-          message: 'Auth failed',
-        });
-      }
-      const token = jwt.sign(
-        { email: fetchedUser.email, userId: fetchedUser._id },
-        'secret_this_should_be_longer',
-        {expiresIn: "1h"},
-      );
-      res.status(200).json({
-        token: token,
-        expiresIn: 3600,
-        userId: fetchedUser._id,
-      });
-    })
-    .catch(err => {
-      return res.status(401).json({
-        message: 'Invalid auth credentials!',
-      });
-    });
-});
+router.post("/login", UserController.userLogin);
 
 module.exports = router;
