+++ src/app/angular-material.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import {
+  MatInputModule,
+  MatCardModule,
+  MatButtonModule,
+  MatToolbarModule,
+  MatExpansionModule,
+  MatProgressSpinnerModule,
+  MatPaginatorModule,
+  MatDialogModule,
+} from '@angular/material';
+
+@NgModule({
+  exports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatToolbarModule,
+    MatExpansionModule,
+    MatProgressSpinnerModule,
+    MatPaginatorModule,
+    MatDialogModule,
+  ]
+})
+export class AngularMaterialModule {}
