+++ src/app/auth/auth.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+
+import { LoginComponent } from './login/login/login.component';
+import { SignupComponent } from './signup/signup/signup.component';
+import { AngularMaterialModule } from '../angular-material.module';
+import { CommonModule } from '@angular/common';
+import { AuthRoutingModule } from './auth-routing.module';
+
+@NgModule({
+  declarations: [
+    LoginComponent,
+    SignupComponent,
+  ],
+  imports: [
+    AngularMaterialModule,
+    CommonModule,
+    FormsModule,
+    AuthRoutingModule
+  ]
+})
+export class AuthModule {}
