+++ src/app/posts/post-list/post-list.component.ts
@@ -9,7 +9,7 @@ import { AuthService } from 'src/app/auth/auth.service';
 @Component({
   selector: 'app-post-list',
   templateUrl: './post-list.component.html',
-  styleUrls: ['./post-list.component.scss']
+  styleUrls: ['./post-list.component.scss'],
 })
 
 export class PostListComponent implements OnInit, OnDestroy {
