+++ src/app/posts/post.service.ts
@@ -6,6 +6,10 @@ import { map } from 'rxjs/operators';
 import { Post } from './post.model';
 import { Router } from '@angular/router';
 
+import { environment } from '../../environments/environment';
+
+const BACKEND_URL = environment.apiUrl + '/posts/';
+
 @Injectable({providedIn: 'root'})
 export class PostService {
   private posts: Post[] = [];
@@ -15,7 +19,7 @@ export class PostService {
 
   getPosts(postsPerPage: number, currentPage: number) {
     const queryParams = `?pagesize=${postsPerPage}&page=${currentPage}`;
-    this.http.get<{ message: string, posts: any, maxPosts: number }>('http://localhost:3000/api/posts' + queryParams)
+    this.http.get<{ message: string, posts: any, maxPosts: number }>(BACKEND_URL + queryParams)
     .pipe(map((postData) => {
       return {
         posts: postData.posts.map(post => {
@@ -51,7 +55,7 @@ export class PostService {
       imagePath: string,
       creator: string,
     }>(
-      'http://localhost:3000/api/posts/' + id
+      BACKEND_URL + id
     );
   }
 
@@ -60,7 +64,7 @@ export class PostService {
     postData.append('title', title);
     postData.append('content', content);
     postData.append('image', image, title);
-    this.http.post<{message: string, post: Post}>('http://localhost:3000/api/posts', postData)
+    this.http.post<{message: string, post: Post}>(BACKEND_URL, postData)
     .subscribe((responseData) => {
       this.router.navigate(['/']);
     });
@@ -83,13 +87,13 @@ export class PostService {
         creator: null
       };
     }
-    this.http.put('http://localhost:3000/api/posts/' + id, postData)
+    this.http.put(BACKEND_URL + id, postData)
     .subscribe(response => {
       this.router.navigate(['/']);
     });
   }
 
   deletePost(postId: string) {
-    return this.http.delete('http://localhost:3000/api/posts/' + postId);
+    return this.http.delete(BACKEND_URL + postId);
   }
 }
