+++ src/app/posts/posts.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostListComponent } from './post-list/post-list.component';
+import { AngularMaterialModule } from '../angular-material.module';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
+
+@NgModule({
+  declarations: [
+    PostCreateComponent,
+    PostListComponent,
+  ],
+  imports: [
+    ReactiveFormsModule,
+    AngularMaterialModule,
+    CommonModule,
+    RouterModule,
+  ]
+})
+export class PostsModule {}
